Implemented functionality to import data from an Excel sheet into a database table. Additionally, created a new table based on the headers present in the Excel sheet. Completed task as per requirements.
