--- import/excel_project/excel_app/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class ExcelAppConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'excel_app'
