--- import/excel_project/excel_app/models.py
@@ -1,11 +1,0 @@
-from django.db import models
-
-class ImportedData(models.Model):
-    name = models.CharField(max_length=100)
-    email = models.EmailField()
-    position = models.CharField(max_length=100)
-    mobile = models.CharField(max_length=100)
-    action = models.CharField(max_length=100)
-
-    def __str__(self):
-        return self.name
