--- import/excel_project/excel_app/urls.py
@@ -1,6 +1,0 @@
-from django.urls import path
-from . import views
-
-urlpatterns = [
-    path('', views.upload_file, name='upload_file'),
-]
