+++ webdem/urls.py
@@ -3,4 +3,5 @@ from webdem import views
 
 urlpatterns = [
     path('',views.index , name="index"),
+    path('upload/',views.upload_file , name="upload_file"),
 ]
